removed opening move code
not needed, asked at the tute

+++ starter_code/Board.cpp
@@ -58,7 +58,7 @@ bool Board::placeTile(Tile &tile, char rowInput, int col) {
 
             int positions = 0;
 
-            for (int i = 0; i < shapes.size(); ++i) {
+            for (int i = 0; i < (int) shapes.size(); ++i) {
                 if (shapesColours[i] != 0) {
                     positions++;
                 }
